+++ examples/tokio-mqtt-connector-demo/src/main.rs
@@ -158,7 +158,7 @@ async fn main() -> DbResult<()> {
             .source(indoor_temp_producer)
             .tap(temperature_logger)
             .link_to(SensorKey::TempIndoor.link_address().unwrap())
-            .with_serializer(|temp: &Temperature| Ok(temp.to_json_vec()))
+            .with_serializer_raw(|temp: &Temperature| Ok(temp.to_json_vec()))
             .finish();
     });
 
@@ -167,7 +167,7 @@ async fn main() -> DbResult<()> {
             .source(outdoor_temp_producer)
             .tap(temperature_logger)
             .link_to(SensorKey::TempOutdoor.link_address().unwrap())
-            .with_serializer(|temp: &Temperature| Ok(temp.to_json_vec()))
+            .with_serializer_raw(|temp: &Temperature| Ok(temp.to_json_vec()))
             .finish();
     });
 
@@ -176,7 +176,7 @@ async fn main() -> DbResult<()> {
             .source(server_room_temp_producer)
             .tap(temperature_logger)
             .link_to(SensorKey::TempServerRoom.link_address().unwrap())
-            .with_serializer(|temp: &Temperature| Ok(temp.to_json_vec()))
+            .with_serializer_raw(|temp: &Temperature| Ok(temp.to_json_vec()))
             .finish();
     });
 
@@ -185,7 +185,7 @@ async fn main() -> DbResult<()> {
         reg.buffer(BufferCfg::SpmcRing { capacity: 10 })
             .tap(command_consumer)
             .link_from(CommandKey::TempIndoor.link_address().unwrap())
-            .with_deserializer(|_ctx, data: &[u8]| TemperatureCommand::from_json(data))
+            .with_deserializer_raw(|data: &[u8]| TemperatureCommand::from_json(data))
             .finish();
     });
 
@@ -193,7 +193,7 @@ async fn main() -> DbResult<()> {
         reg.buffer(BufferCfg::SpmcRing { capacity: 10 })
             .tap(command_consumer)
             .link_from(CommandKey::TempOutdoor.link_address().unwrap())
-            .with_deserializer(|_ctx, data: &[u8]| TemperatureCommand::from_json(data))
+            .with_deserializer_raw(|data: &[u8]| TemperatureCommand::from_json(data))
             .finish();
     });